pyproject.toml
[project]
name = "traffic-zero"
version = "0.1.0"
description = "Adaptive multi-agent traffic lights system"
readme = "README.md"
dependencies = [
    "numpy",
    "pandas",
    "sumolib",
    "traci",
    "seaborn",
    "scipy",
    "networkx",
    "torch",
    "gymnasium",
    "ray[rllib]; platform_system != 'Windows'",
    "tensorboard",
    "traci>=1.19.0",
    "opencv-python",
    "flake8-bugbear>=24.12.12",
    "flake8-comprehensions>=3.17.0",
    "flake8-docstrings>=1.7.0",
    "flake8>=7.3.0",
    "black[jupyter]>=25.9.0",
    "pre-commit>=4.3.0",
    "pettingzoo>=1.25.0",
    "pygame>=2.6.1",
    "mpe2>=0.0.1",
    "dash>=3.2.0",
    "carla>=0.9.16",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "pip>=25.3",
    "ipykernel>=7.1.0",
    "pyarrow>=21.0.0",
    "tabulate>=0.9.0",
    "scikit-learn>=1.7.2",
    "dython>=0.7.11",
    "tqdm>=4.67.1",
]
requires-python = "==3.12"
license = {text = "MIT"}

[tool.pyright]
include = ["**.py"]
exclude = [
    "**/site-packages/**",
    "**/.venv/**",
    "**/node_modules/**",
    "**/__pycache__/**"
]
reportMissingImports = false

[tool.black]
line-length = 80
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(\.
    | \.venv
    | build
    | dist
    | .eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
)/
'''

[tool.isort]
profile = "black"
line_length = 80
known_first_party = ["traffic_zero"]

[tool.flake8]
max-line-length = 80
extend-ignore = [
    "E203",  # whitespace before ':', conflicts with black
    "W503",  # line break before binary operator, conflicts with black
    "D107",  # missing docstring in __init__
    "D202",  # No blank lines allowed after function docstring
    "D401",  # First word of the docstring should be properly capitalized
    "B008",  # Do not perform function calls in argument defaults
]
